pull LRU crawler out into its own file.
~600 lines gone from items.c makes it a lot more manageable.

this change is almost purely moving code around and renaming functions. very
little logic has changed.
